MSAPI-101 | Add email validator
